refactor(ui): 重构代码结构，调整文件路径和导入引用
- 将多个 UI 组件和模块从 src/ui 目录移动到 src/ui/views 目录
- 更新所有相关文件中的导入路径，确保引用正确
- 修改多个类型导入路径，从 session-types 改为 session/types
- 统一调整 hooks 和核心模块的导入路径，修正导入命名和位置
- 维护代码一致性，避免因路径更改导致的引用错误
- 重命名 session.ts 为 session/index.ts，并修改相关导入
- 更新测试用例中的类型导入路径以匹配新的文件结构
- 将多处核心或公共模块导入路径调整为模块根目录的对应位置
