+++ src/tests/promptInputKeys.test.ts
@@ -14,20 +14,19 @@ import {
   getPromptCursorPlacement,
   getPromptReturnKeyAction,
   isClearImageAttachmentsShortcut,
-  parseTerminalInput,
   removeCurrentSlashToken,
   toggleSkillSelection,
   renderBufferWithCursor,
   buildInitPromptSubmission,
   buildPromptDraftFromSessionMessage,
-  dispatchTerminalInput,
   disableTerminalExtendedKeys,
   enableTerminalExtendedKeys,
   EMPTY_BUFFER,
   insertText,
   backspace,
 } from "../ui";
-import type { SessionMessage, SkillInfo } from "../session-types";
+import type { SessionMessage, SkillInfo } from "../session/types";
+import { dispatchTerminalInput, parseTerminalInput } from "../ui/hooks";
 
 function collectDispatchedInput(data: string) {
   const events: ReturnType<typeof parseTerminalInput>[] = [];
