+++ src/tests/session.test.ts
@@ -5,7 +5,7 @@ import * as fs from "fs";
 import * as os from "os";
 import * as path from "path";
 import { GitFileHistory } from "../common/runtime/file-history";
-import { type SessionMessage } from "../session-types";
+import { type SessionMessage } from "../session/types";
 import { SessionManager } from "../session";
 
 const originalFetch = globalThis.fetch;
