+++ src/ui/views/PermissionPrompt.tsx
@@ -1,8 +1,8 @@
 import React, { useEffect, useMemo, useState } from "react";
 import { Box, Text } from "ink";
-import { useTerminalInput } from "./hooks";
-import type { AskPermissionRequest, AskPermissionScope, UserToolPermission } from "../common/permissions";
-import type { PermissionScope } from "../settings";
+import { useTerminalInput } from "../hooks";
+import type { AskPermissionRequest, AskPermissionScope, UserToolPermission } from "../../common/permissions";
+import type { PermissionScope } from "../../settings";
 
 export type PermissionPromptResult = {
   permissions: UserToolPermission[];
