+++ src/ui/views/ProcessStdoutView.tsx
@@ -1,8 +1,8 @@
 import React, { useEffect, useMemo, useRef, useState } from "react";
 import { Box, Text } from "ink";
-import { BASH_TIMEOUT_DECREMENT_MS, BASH_TIMEOUT_INCREMENT_MS } from "../common/system/bash-timeout";
-import type { BashTimeoutAdjustment, SessionEntry, SessionProcessEntry } from "../session-types";
-import { useTerminalInput } from "./hooks";
+import { BASH_TIMEOUT_DECREMENT_MS, BASH_TIMEOUT_INCREMENT_MS } from "../../common/system/bash-timeout";
+import type { BashTimeoutAdjustment, SessionEntry, SessionProcessEntry } from "../../session/types";
+import { useTerminalInput } from "../hooks";
 
 type RunningProcesses = SessionEntry["processes"];
