+++ src/ui/views/WelcomeScreen.tsx
@@ -2,12 +2,12 @@ import React, { useMemo, useState } from "react";
 import { Box, Text } from "ink";
 import * as os from "node:os";
 import path from "node:path";
-import type { SkillInfo } from "../session-types";
-import type { ResolvedDeepcodingSettings } from "../settings";
-import { buildSlashCommands, BUILTIN_SLASH_COMMANDS, formatSlashCommandDescription } from "./core/slashCommands";
+import type { SkillInfo } from "../../session/types";
+import type { ResolvedDeepcodingSettings } from "../../settings";
+import { buildSlashCommands, BUILTIN_SLASH_COMMANDS, formatSlashCommandDescription } from "../core/slashCommands";
 import { ThemedGradient } from "./ThemedGradient";
-import { AsciiLogo } from "./AsciiArt";
-import { useAppContext } from "./contexts";
+import { AsciiLogo } from "../AsciiArt";
+import { useAppContext } from "../contexts";
 
 type WelcomeScreenProps = {
   projectRoot: string;
